["latex", "verbatim"] make portable by using cat
... and the BLOB construct instead of echo.

The echo(1) function is usually shell built in. This in itself
is a reasonalbe design decission, as there is no need to start
an external program just to join a couple of strings by spaces.
Unfortunately, however, shells don't agree on the semantics of
echo, in particular about what the correct invocation is to
output a literal backslash; this can be seen from the following
invovations.

  ~>dash -c 'echo '\''\\'\'''
  \
  ~>bash -c 'echo '\''\\'\'''
  \\
  ~>

Now, both of those shells can act as the, supposedly POSIX-compliant
sh. Work around these incompatibilities by generating BLOBs for
the \begin/\end{verbatim} literal and join them with cat(1).